refactor: move tests out to own file
