+++ tests/test_package_toml.py
@@ -1,7 +1,6 @@
-import os
 from package import get_build_system_from_pyproject_toml, BuildPlanManager
 from pytest import raises
-from unittest.mock import MagicMock, Mock
+from unittest.mock import Mock
 
 
 def test_get_build_system_from_pyproject_toml_inexistent():
@@ -40,47 +39,3 @@ def test_get_build_system_from_pyproject_toml_poetry():
         )
         == "poetry"
     )
-
-
-def test_zip_source_path_sh_work_dir():
-    zs = Mock()
-    zs.write_dirs = MagicMock()
-
-    bpm = BuildPlanManager(args=Mock())
-    bpm._zip_write_with_filter = MagicMock()
-
-    bpm.execute(
-        build_plan=[
-            ["sh", ".", "cd $(mktemp -d)\n echo pip install"],
-            ["zip:embedded", ".", "./python"],
-        ],
-        zip_stream=zs,
-        query=None,
-    )
-
-    zs.write_dirs.assert_called_once()
-
-    zip_source_path = zs.write_dirs.call_args_list[0][0][0]
-    assert zip_source_path != f"{os.getcwd()}"
-
-
-def test_zip_source_path():
-    zs = Mock()
-    zs.write_dirs = MagicMock()
-
-    bpm = BuildPlanManager(args=Mock())
-    bpm._zip_write_with_filter = MagicMock()
-
-    bpm.execute(
-        build_plan=[
-            ["sh", ".", "echo pip install"],
-            ["zip:embedded", ".", "./python"],
-        ],
-        zip_stream=zs,
-        query=None,
-    )
-
-    zs.write_dirs.assert_called_once()
-
-    zip_source_path = zs.write_dirs.call_args_list[0][0][0]
-    assert zip_source_path == f"{os.getcwd()}"
